# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API for UNICORN application",
        "title": "UNICORN API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/unicorn": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unicorn"
                ],
                "summary": "Create Unicorn Process.",
                "operationId": "createUnicorn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "/api/create-unicorn?amount=...",
                        "name": "amount",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/domain.UnicornProcess"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Error in process"
                    }
                }
            }
        },
        "/api/unicorn/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unicorn"
                ],
                "summary": "Search Unicorn Process.",
                "operationId": "getUnicorn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Unicorn"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Error in process"
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Unicorn": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.UnicornProcess": {
            "type": "object",
            "properties": {
                "requestID": {
                    "type": "string"
                }
            }
        }
    }
}